全画面表示時のコードを切り出す
